simulator/requirements.txt
# This file should include the union of all the Python dependencies needed for all the models.
msal-extensions==0.1.3
numpy>=1.15.1
python-dotenv==0.13.0
bonsai-cli==1.0.0
microsoft-bonsai-api==0.1.1
pyglet==1.5.0
pandas==1.2.2
gym==0.18.0
Box2D==2.3.2
requests==2.22.0
